+++ CHANGELOG.md
@@ -18,7 +18,9 @@
 
 #### Core
 
-- overhaul termination handler invocation
+- **Breaking Change**: overhaul ``TERM_HOOK`` invocation
+    - change ``TERM_HOOK`` interface, now do not pass extra information (exit status, termination kind)
+    - also remove ``ON_ASSERT``, ``ON_ERR``, ``ON_EXIT`` constants
 - send ``SIGHUP`` to manged jobs before process termination (even if subshell)
 
 ### Fixed